[cli-tests] Refactor multipass cli spawn
